Move navbar files to their own folder
Make navbar a component of a layout
