+++ frontend/src/routes/Router.tsx
@@ -1,8 +1,9 @@
 import { AuthenticationGuard } from "components/common/AuthenticationGuard";
+import Landing from "components/modules/Landing/Landing";
+import Layout from "layouts/Layout";
 import LoginForm from "components/modules/Login/Login";
 import RegisterForm from "components/modules/Registeration/Registration";
-import Landing from "components/modules/Landing/Landing";
-import NavigationBar from "layouts/NavigationBar";
+
 import React from "react";
 import {
   createBrowserRouter,
@@ -12,9 +13,10 @@ import {
 
 const routes = createRoutesFromElements(
   <Route>
-    {/* Unauthenticated landing: */}
-    <Route index element={<Landing />} />
-    <Route path="/" element={<NavigationBar />}>
+    {/* TODO: Different navbar/layout for the unauthenticated. */}
+    {/* Unauthenticated stuff: */}
+    <Route path="/" element={<Layout />}>
+      <Route index element={<Landing />} />
       <Route
         element={
           <AuthenticationGuard
@@ -36,10 +38,12 @@ const routes = createRoutesFromElements(
         <Route path="register" element={<RegisterForm />} />
       </Route>
     </Route>
-    {/* Protect all other routes based on authentication.
-        I.e., put all stuff in here: */}
+
+    {/* Protect all other routes with authentication.
+        I.e., put all stuff in here:
+    */}
     <Route element={<AuthenticationGuard />}>
-      <Route path="/" element={<NavigationBar />}></Route>
+      <Route path="/" element={<Layout />}></Route>
     </Route>
   </Route>
 );
